Changed the way business list was displayed
Chose dictionary because that way it gives us the restaurant and category in one item, as opposed to a list, which would give the restaurant and category in 2 separate items. So that when we add the breakfast list to the meal itinerary data frame, restaurant name and category are together as opposed to two separate items.