stuck on business list
converted business list display to return a dictionary of restaurant name and category.

however, stuck on how to individually append each dictionary key value pair to breakfast dict so that we can add breakfast dict to the meal itinerary df. Tried appending the breakfast_dict to the breakfast_list so we can put the breakfast_list in the DF, but it only returns the same restaurant/category pair instead of multiple different res/cat pairs