Working on breakfast list (#5)
* Capturing price errors

captured errors for price limit inputs (if price limit input is numeric, alphabetic, or has more $ signs than 4)

* Edited display of breakfast list

Displayed breakfast list in the format of Restaurant Name | Category

* Edited breakfast parameters

breakfast parameters limit:
instead of returning 50 restaurants, it now returns the amount of restaurants same as our number of vacation days. since this is just the breakfast list, it will return to us 1 bfast per day, according to our number of vacay days

* Tried inserting images (doesn't work yet)

-Got image URL
-tried importing urllib.request, downloading PIL and importing Image
-using Image and urllib.request.urlopen to display image
-image doesn't show but it returns an HTML code (wondering if this would show up on a web app??)
-will keep working on this after

* Changed the way business list was displayed

Chose dictionary because that way it gives us the restaurant and category in one item, as opposed to a list, which would give the restaurant and category in 2 separate items. So that when we add the breakfast list to the meal itinerary data frame, restaurant name and category are together as opposed to two separate items.

* stuck on business list

converted business list display to return a dictionary of restaurant name and category.

however, stuck on how to individually append each dictionary key value pair to breakfast dict so that we can add breakfast dict to the meal itinerary df. Tried appending the breakfast_dict to the breakfast_list so we can put the breakfast_list in the DF, but it only returns the same restaurant/category pair instead of multiple different res/cat pairs

Co-authored-by: bz150 <77802960+bz150@users.noreply.github.com>